requirements.txt
streamlit
torch
transformers
pandas
scikit-learn
numpy
tqdm
Pillow
